refactor: merge BytecodeInstructions into Instructions
