+++ main/src/ca/uwaterloo/flix/language/phase/jvm/GenAnonymousClasses.scala
@@ -20,7 +20,7 @@ import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.language.ast.shared.{JConstructor, JMethod}
 import ca.uwaterloo.flix.language.ast.{AtomicOp, SimpleType}
 import ca.uwaterloo.flix.language.ast.JvmAst.*
-import ca.uwaterloo.flix.language.phase.jvm.BytecodeInstructions.*
+import ca.uwaterloo.flix.language.phase.jvm.Instructions.*
 import ca.uwaterloo.flix.language.phase.jvm.ClassMaker.Final.{IsFinal, NotFinal}
 import ca.uwaterloo.flix.language.phase.jvm.ClassMaker.Visibility.IsPublic
 import ca.uwaterloo.flix.language.phase.jvm.ClassMaker.Volatility.NotVolatile
@@ -98,7 +98,7 @@ object GenAnonymousClasses {
   }
 
   private def constructorIns(superClass: ClassDesc)(implicit mv: MethodVisitor): Unit = {
-    import BytecodeInstructions.*
+    import Instructions.*
     ALOAD(0)
     INVOKESPECIAL(ClassMaker.ConstructorMethod(superClass, Nil))
     RETURN()
@@ -106,7 +106,7 @@ object GenAnonymousClasses {
 
   /** Creates constructor bytecode that forwards parameters directly to the super constructor. */
   private def constructorInsWithSuperCall(superClass: ClassDesc, constructor: JConstructor)(implicit mv: MethodVisitor): Unit = {
-    import BytecodeInstructions.*
+    import Instructions.*
     // ALOAD 0 (this)
     thisLoad()
     // Load each <init> parameter (starting at slot 1)
